Refs #105.
Don't delete recordedrating, recordedprogram, and recordedcredits table
entries for a program when we delete the recorded table entry unless we're
deleting the last recording with that same progstart.  Previously if you
had 2 recordings covering the same program (such as when you stop and
restart a recording for some reason), the three tables above would be cleaned
up when the first part was deleted.  Should Close #105.  If it doesn't, then
the ticket can be reopened if someone has more detailed information.



git-svn-id: http://svn.mythtv.org/svn/trunk@7986 7dbf422c-18fa-0310-86e9-fd20926502f2

+++ mythtv/programs/mythbackend/mainserver.cpp
@@ -1391,9 +1391,8 @@ void MainServer::DoDeleteThread(DeleteStruct *ds)
     query.bindValue(":CHANID", ds->chanid);
     query.bindValue(":TITLE", ds->title.utf8());
     query.bindValue(":STARTTIME", ds->recstartts);
-    query.exec();
 
-    if (!query.isActive())
+    if (!query.exec() || !query.isActive())
     {
         MythContext::DBError("Recorded program deletion", query);
         gContext->LogEntry("mythbackend", LP_ERROR, "Delete Recording",
@@ -1401,32 +1400,42 @@ void MainServer::DoDeleteThread(DeleteStruct *ds)
                                    .arg(logInfo));
     }
 
-    query.prepare("DELETE FROM recordedrating "
-                  "WHERE chanid = :CHANID AND starttime = :STARTTIME;");
+    query.prepare("SELECT COUNT(*) FROM recorded "
+                  "WHERE chanid = :CHANID AND progstart = :PROGSTART;");
     query.bindValue(":CHANID", ds->chanid);
-    query.bindValue(":STARTTIME", pginfo->startts);
-    query.exec();
-    if (!query.exec() || !query.isActive())
-        MythContext::DBError("Recorded program delete recordedrating",
-                             query);
+    query.bindValue(":PROGSTART", pginfo->startts);
+    if (!query.exec() || !query.isActive() || !query.next())
+    {
+        MythContext::DBError("Recorded program parts left query", query);
+    }
+    else if (query.value(0).toInt() == 0)
+    {
+        // No more recordings that are part of this Program so we can
+        // cleanup these three tables
+        query.prepare("DELETE FROM recordedrating "
+                      "WHERE chanid = :CHANID AND starttime = :STARTTIME;");
+        query.bindValue(":CHANID", ds->chanid);
+        query.bindValue(":STARTTIME", pginfo->startts);
+        if (!query.exec() || !query.isActive())
+            MythContext::DBError("Recorded program delete recordedrating",
+                                 query);
 
-    query.prepare("DELETE FROM recordedprogram "
-                  "WHERE chanid = :CHANID AND starttime = :STARTTIME;");
-    query.bindValue(":CHANID", ds->chanid);
-    query.bindValue(":STARTTIME", pginfo->startts);
-    query.exec();
-    if (!query.exec() || !query.isActive())
-        MythContext::DBError("Recorded program delete recordedprogram",
-                             query);
+        query.prepare("DELETE FROM recordedprogram "
+                      "WHERE chanid = :CHANID AND starttime = :STARTTIME;");
+        query.bindValue(":CHANID", ds->chanid);
+        query.bindValue(":STARTTIME", pginfo->startts);
+        if (!query.exec() || !query.isActive())
+            MythContext::DBError("Recorded program delete recordedprogram",
+                                 query);
 
-    query.prepare("DELETE FROM recordedcredits "
-                  "WHERE chanid = :CHANID AND starttime = :STARTTIME;");
-    query.bindValue(":CHANID", ds->chanid);
-    query.bindValue(":STARTTIME", pginfo->startts);
-    query.exec();
-    if (!query.exec() || !query.isActive())
-        MythContext::DBError("Recorded program delete recordedcredits",
-                             query);
+        query.prepare("DELETE FROM recordedcredits "
+                      "WHERE chanid = :CHANID AND starttime = :STARTTIME;");
+        query.bindValue(":CHANID", ds->chanid);
+        query.bindValue(":STARTTIME", pginfo->startts);
+        if (!query.exec() || !query.isActive())
+            MythContext::DBError("Recorded program delete recordedcredits",
+                                 query);
+    }
 
     sleep(1);
